+++ src/backend/resolvers/mutations/collection/uploadSheet/tests/validateSheetData/mockData/input/businessObjValidationInput/input1/index.js
@@ -1,0 +1,9 @@
+const BO_INPUT_1_dataAndSkippedRows = require('./data-and-skipped-rows')
+const BO_INPUT_1_mockDb = require('./mock-db')
+const BO_INPUT_1_sheetConfig = require('./sheet-config')
+
+module.exports = {
+  BO_INPUT_1_dataAndSkippedRows,
+  BO_INPUT_1_mockDb,
+  BO_INPUT_1_sheetConfig,
+}