+++ src/frontend/api/queries.js
@@ -547,10 +547,7 @@ export const GET_WORKBOOKS = gql`
           name
           type
           oneOf
-          businessObjRef {
-            _id
-            fieldId
-          }
+          businessObjRef
         }
       }
     }